fix(test): 更新哈希表测试接口并修复信号注销逻辑
- 将 eh_hashtbl_node_new 和 eh_hashtbl_node_new_with_string 接口替换为带 refresh 后缀的新接口，确保节点创建行为一致。
- 同时修正信号与槽的注销流程，使用新的 eh_signal_slot_disconnect和 eh_signal_slot_clean 接口替代旧有函数，提升代码健壮性。
- 移除未使用的测试可执行文件定义，优化内部调度计数器类型。

+++ src/include/eh_interior.h
@@ -37,7 +37,7 @@ struct eh{
     struct      eh_task                  *main_task;                                            /* 系统栈任务 */
     struct      eh_module                *eh_init_fini_array;
     long                                 eh_init_fini_array_len;
-    unsigned    long                     dispatch_cnt;                                          /* 调度次数 */
+    unsigned int                         dispatch_cnt;                                          /* 调度次数 */
     eh_clock_t                           idle_time;                                             /* 空闲任务占用的时间片 */
 };
